C qube release qa

+++ tests/src/CRC/check_districtwise_records.py
@@ -28,6 +28,8 @@ class test_crc_report_districtwise():
         for x in range(1, len(select_district.options)):
             select_district.select_by_index(x)
             self.cal.page_loading(self.driver)
+            value = self.driver.find_element_by_name('myDistrict').get_attribute('value')
+            distval=value[3:]+'_'
             nodata = self.driver.find_element_by_id("errMsg").text
             if nodata == "No data found":
                 print(select_district.options[x].text, "no data found!")
@@ -35,31 +37,35 @@ class test_crc_report_districtwise():
             else:
                 self.driver.find_element_by_id(Data.Download).click()
                 time.sleep(3)
-                self.filename = p.get_download_dir() + '/' + self.fname.crc_districtwise()
-                with open(self.filename) as fin:
-                    csv_reader = csv.reader(fin, delimiter=',')
-                    header = next(csv_reader)
-                    tschools = 0
-                    vsts = 0
-                    vstd = 0
-                    for row in csv.reader(fin):
-                        tschools += int(row[0])
-                        vsts +=int(row[2])
-                        vstd +=int(row[1])
-                    totalschools = self.driver.find_element_by_id("schools").text
-                    visited = self.driver.find_element_by_id("visited").text
-                    visits = self.driver.find_element_by_id("visits").text
-                    tsc = re.sub('\D',"",totalschools)
-                    vs = re.sub('\D',"",visits)
-                    vd= re.sub('\D', "",visited)
-                    if int(tsc) != tschools:
-                        print(select_district.options[x].text, ":", "total no of schools  :",tschools, int(tsc),"records are mismatch found")
-                    if int(vs) != vsts:
-                        print(select_district.options[x].text, ":", "total no of visits  :",int(vs) , vsts,"records are mismatch found")
-                    if int(vd) != vstd:
-                        print(select_district.options[x].text, ":", "total no of visits  :", int(vd), vstd,
-                              "records are mismatch found")
+                self.filename = p.get_download_dir() + '/' + self.fname.crc_districtwise()+distval.strip()+self.cal.get_current_date()+'.csv'
+                if os.path.isfile(self.filename) != True:
+                    print(select_district.options[x].text,'csv file is not downloaded')
+                    count = count + 1
+                else:
+                    with open(self.filename) as fin:
+                        csv_reader = csv.reader(fin, delimiter=',')
+                        header = next(csv_reader)
+                        tschools = 0
+                        vsts = 0
+                        vstd = 0
+                        for row in csv.reader(fin):
+                            tschools += int(row[0])
+                            vsts +=int(row[2])
+                            vstd +=int(row[1])
+                        totalschools = self.driver.find_element_by_id("schools").text
+                        visited = self.driver.find_element_by_id("visited").text
+                        visits = self.driver.find_element_by_id("visits").text
+                        tsc = re.sub('\D',"",totalschools)
+                        vs = re.sub('\D',"",visits)
+                        vd= re.sub('\D', "",visited)
+                        if int(tsc) != tschools:
+                            print(select_district.options[x].text, ":", "total no of schools  :",tschools, int(tsc),"records are mismatch found")
+                        if int(vs) != vsts:
+                            print(select_district.options[x].text, ":", "total no of visits  :",int(vs) , vsts,"records are mismatch found")
+                        if int(vd) != vstd:
+                            print(select_district.options[x].text, ":", "total no of visits  :", int(vd), vstd,
+                                  "records are mismatch found")
 
 
-                os.remove(self.filename)
+                    os.remove(self.filename)
         return count
