+++ tests/src/Teacher_Attendance/download_clusterwise_csv.py
@@ -2,6 +2,7 @@ import os
 import time
 
 from Data.parameters import Data
+from filenames import file_extention
 from get_dir import pwd
 from reuse_func import GetData
 
@@ -15,15 +16,21 @@ class ClusterwiseCsv():
 
     def click_download_icon_of_clusters(self):
         cal = GetData()
+        files = file_extention()
         cal.click_on_state(self.driver)
         cal.page_loading(self.driver)
+        self.year,self.month = cal.get_student_month_and_year_values()
         self.driver.find_element_by_id(Data.SAR_Clusters_btn).click()
         cal.page_loading(self.driver)
+        time.sleep(5)
         self.driver.find_element_by_id(Data.Download).click()
         time.sleep(5)
         p = pwd()
-        self.filename = p.get_download_dir() + "/Cluster_wise_report_" + self.month + "_" + self.year + ".csv"
-        return os.path.isfile(self.filename)
-
-    def remove_csv(self):
-        os.remove(self.filename)
+        count = 0
+        self.filename = p.get_download_dir() +'/'+files.teacher_cluster_download()+self.month+"_"+self.year+'_'+cal.get_current_date()+".csv"
+        print(self.filename)
+        if os.path.isfile(self.filename) !=True:
+            count = count + 1
+        else:
+            os.remove(self.filename)
+        return count
